chore: point the generated SubmitFunction alias at $app/forms
