+++ packages/kit/src/core/sync/write_types/test/simple-page-server-only/tsconfig.json
@@ -8,6 +8,7 @@
 		"moduleResolution": "bundler",
 		"paths": {
 			"@sveltejs/kit": ["../../../../../exports/public"],
+			"$app/forms": ["../../../../../runtime/app/forms/types"],
 			"types": ["../../../../../types/internal"],
 			"$app/types": ["../../../../../types/ambient.d.ts"]
 		},
